causal_graph_discovery_with_llms/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
contourpy==1.3.0
cycler==0.12.1
distro==1.9.0
fonttools==4.53.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jiter==0.5.0
kiwisolver==1.4.7
matplotlib==3.9.2
networkx==3.3
numpy==2.0.1
openai==1.40.8
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
sniffio==1.3.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
